Adjusting WARNING period LED visual
Yanking some duplicate logic out of the main loop and creating a
SHUTDOWN_WARNING_PERIOD boolean to keep track of things. Clarified the
names of some constants.  Restored the outside "if" statement of "if
(fireplaceOn)" into the countDownTimer loop as I believe it was correct
in the first place and not causing the weird issues I ran into before.